Allow Documents to be created from string [skip ci]
