+++ src/Document.ts
@@ -2,7 +2,7 @@ import { assignIn, cloneDeep, get as getIn, set as setIn } from 'lodash';
 import { TransformDataError } from './errors';
 import ForeignKeyDbTransformer from './ForeignKeyDbTransformer';
 import type Schema from './Schema';
-import type { GenericObject, MvRecord } from './types';
+import type { DbServerDelimiters, GenericObject, MvRecord } from './types';
 
 // #region Types
 export interface DocumentConstructorOptions {
@@ -53,6 +53,53 @@ class Document {
 		}
 		// load the data passed to constructor into document instance
 		assignIn(this, data);
+	}
+
+	/** Create a new Document instances from a record string */
+	public static createDocumentFromRecordString(
+		schema: Schema,
+		recordString: string,
+		dbServerDelimiters: DbServerDelimiters,
+	): Document {
+		const record = Document.convertMvStringToArray(recordString, dbServerDelimiters);
+
+		return new Document(schema, { record });
+	}
+
+	/** Convert a multivalue string to an array */
+	protected static convertMvStringToArray(
+		recordString: string,
+		dbServerDelimiters: DbServerDelimiters,
+	): MvRecord {
+		const { am, vm, svm } = dbServerDelimiters;
+		const record: MvRecord =
+			recordString === ''
+				? []
+				: recordString.split(am).map((attribute) => {
+						if (attribute === '') {
+							return null;
+						}
+
+						const attributeArray = attribute.split(vm);
+						if (attributeArray.length === 1) {
+							return attribute;
+						}
+
+						return attributeArray.map((value) => {
+							if (value === '') {
+								return null;
+							}
+
+							const valueArray = value.split(svm);
+							if (valueArray.length === 1) {
+								return value;
+							}
+
+							return valueArray.map((subvalue) => (subvalue === '' ? null : subvalue));
+						});
+				  });
+
+		return record;
 	}
 
 	/** Transform document structure to multivalue array structure */
