+++ src/utils/index.ts
@@ -1,3 +1,2 @@
-export { default as convertMvStringToArray } from './convertMvStringToArray';
 export { default as dummyLogger } from './dummyLogger';
 export { default as ensureArray } from './ensureArray';
